db: update datatype of FK relation
In the revision table, the datatype of rev_id was changed from int
to bigint. mathindex_revision_id references that field
and needs to be updated as well.

* Use a newer method to get $dbr

Bug: T365445

+++ db/math_mlp.sql
@@ -3,7 +3,7 @@ CREATE TABLE /*_*/math_mlp (
   user_id int(10) unsigned NOT NULL,
   step TINYINT NOT NULL,
   anchor varchar(50) DEFAULT NULL,
-  revision_id int(10) unsigned NOT NULL,
+  revision_id bigint(20) unsigned NOT NULL,
   json_data MEDIUMTEXT,
   comment varchar(255) DEFAULT NULL,
   timestamp TIMESTAMP DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP,
